@@ -16,10 +16,9 @@
         }
     ],
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1",
         "psr/container": "^1.1|^2.0",
         "symfony/deprecation-contracts": "^2.1|^3",
-        "symfony/polyfill-php81": "^1.22",
         "symfony/service-contracts": "^1.1.6|^2.0|^3.0"
     },
     "require-dev": {
