@@ -21,8 +21,8 @@ class RewindableGenerator implements \IteratorAggregate, \Countable
 
     public function __construct(callable $generator, int|callable $count)
     {
-        $this->generator = $generator instanceof \Closure ? $generator : \Closure::fromCallable($generator);
-        $this->count = \is_callable($count) && !$count instanceof \Closure ? \Closure::fromCallable($count) : $count;
+        $this->generator = $generator(...);
+        $this->count = \is_int($count) ? $count : $count(...);
     }
 
     public function getIterator(): \Traversable
@@ -34,7 +34,7 @@ class RewindableGenerator implements \IteratorAggregate, \Countable
 
     public function count(): int
     {
-        if (\is_callable($count = $this->count)) {
+        if (!\is_int($count = $this->count)) {
             $this->count = $count();
         }
 
